% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAnnoVizApp.R
\name{GAnnoVizApp}
\alias{GAnnoVizApp}
\title{Launch GAnnoViz Shiny application}
\usage{
GAnnoVizApp(host = NULL, port = NULL, launch.browser = TRUE, ...)
}
\arguments{
\item{host}{Host interface, e.g., \bold{\emph{"0.0.0.0"}} or "127.0.0.1".}

\item{port}{Port number, e.g., \bold{\emph{8888}}.}

\item{launch.browser}{Whether to launch in a browser. (\bold{\emph{TRUE}}).}

\item{...}{Additional parameters passed to \bold{\emph{shiny::runApp}}.}
}
\description{
Launch the \bold{\emph{GAnnoViz}} Shinyapp.
}
\examples{
# GAnnoVizApp()

}
