% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_theme}
\alias{my_theme}
\title{Publication-ready ggplot theme}
\usage{
my_theme()
}
\value{
A \bold{\emph{ggplot theme}} object.
}
\description{
Publication-ready \bold{\emph{ggplot theme}}.
}
\author{
benben-miao
}
\keyword{internal}
