% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrom_snp.R
\name{plot_chrom_snp}
\alias{plot_chrom_snp}
\title{Plot SNP density at chromosome level}
\usage{
plot_chrom_snp(
  fst_file,
  LOG10 = FALSE,
  bin_size = 1e+06,
  density_color = c("#0088ff", "#ff8800", "#ff0000")
)
}
\arguments{
\item{fst_file}{FST slide window file (\bold{\emph{tab-separated}}).}

\item{LOG10}{Whether to transform values by log10 before plotting. (\bold{\emph{FALSE}}).}

\item{bin_size}{Window size (bp) for density calculation. (\bold{\emph{1e6}}).}

\item{density_color}{Color palette for chromosome density. (\bold{\emph{c("#0088ff", "#ff8800", "#ff0000")}}).}
}
\value{
Plot SNP density at chromosome level.
}
\description{
Plot \bold{\emph{SNP density}} at chromosome level.
}
\examples{
# Example data in SlideAnno
fst_file <- system.file(
  "extdata",
  "example.fst",
  package = "SlideAnno")

# Plot SNP density
plot_chrom_snp(
  fst_file = fst_file,
  LOG10 = FALSE,
  bin_size = 1e6,
  density_color = c("#0088ff", "#ff8800", "#ff0000")
)

}
