% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deg_volcano.R
\name{plot_deg_volcano}
\alias{plot_deg_volcano}
\title{Plot differentially expressed genes (DEGs) volcano}
\usage{
plot_deg_volcano(
  deg_file,
  id_col = "GeneID",
  fc_col = "log2FoldChange",
  sig_col = "padj",
  fc_threshold = 1,
  sig_threshold = 0.05,
  point_size = 2,
  point_alpha = 0.5,
  up_color = "#ff0000",
  down_color = "#008800",
  ns_color = "#888888"
)
}
\arguments{
\item{deg_file}{DEG table from \bold{\emph{DESeq2}} analysis.}

\item{id_col}{Gene IDs column name. (\bold{\emph{"GeneID"}}).}

\item{fc_col}{Log2(fold change) column name. (\bold{\emph{"log2FoldChange"}}).}

\item{sig_col}{Significance column name. (\bold{\emph{"padj"}}, "pvalue").}

\item{fc_threshold}{Absolute log2FC threshold. (\bold{\emph{1}}).}

\item{sig_threshold}{Significance threshold. (\bold{\emph{0.05}}).}

\item{point_size}{Point size. (\bold{\emph{2}}).}

\item{point_alpha}{Point alpha. (\bold{\emph{0.5}}).}

\item{up_color}{Color for up-regulated significant genes. (\bold{\emph{"#ff0000"}}).}

\item{down_color}{Color for down-regulated significant genes. (\bold{\emph{"#008800"}}).}

\item{ns_color}{Color for non-significant genes. (\bold{\emph{"#888888"}}).}
}
\value{
A \bold{\emph{ggplot object}} visualizing DEGs volcano.
}
\description{
Plot differentially expressed genes (DEGs) \bold{\emph{volcano}}.
}
\examples{
# DEG results from DESeq2
deg_file <- system.file(
  "extdata",
  "example.deg",
  package = "GAnnoViz")

# Volcano plot
plot_deg_volcano(
  deg_file = deg_file,
  id_col = "GeneID",
  fc_col = "log2FoldChange",
  sig_col = "padj",
  fc_threshold = 1,
  sig_threshold = 0.05,
  point_size = 2,
  point_alpha = 0.5,
  up_color = "#ff0000",
  down_color = "#008800",
  ns_color = "#888888"
)

}
\author{
benben-miao
}
