% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dmg_exp.R
\name{plot_dmg_exp}
\alias{plot_dmg_exp}
\title{Plot DMGs hyper/hypo along chromosomes}
\usage{
plot_dmg_exp(
  dmr_file,
  orientation = "horizontal",
  chrom_alpha = 0.1,
  chrom_color = "#008888",
  bar_height = 0.8,
  point_size = 1,
  point_alpha = 0.3,
  hyper_color = "#ff0000",
  hypo_color = "#008800",
  mark_style = c("point", "line"),
  line_width = 0.6,
  line_height = 0.8
)
}
\arguments{
\item{dmr_file}{DEG table from \bold{\emph{MethylKit}} analysis.}

\item{orientation}{Coordinate orientation. (\bold{\emph{"horizontal"}}, "vertical").}

\item{chrom_alpha}{Chromosome bar alpha. (\bold{\emph{0.1}}).}

\item{chrom_color}{Chromosome bar color. (\bold{\emph{"#008888"}}).}

\item{bar_height}{Chromosome bar thickness. (\bold{\emph{0.8}}).}

\item{point_size}{Point size. (\bold{\emph{1}}).}

\item{point_alpha}{Point alpha. (\bold{\emph{0.3}}).}

\item{hyper_color}{Color for hyper-methylated. (\bold{\emph{"#ff0000"}}).}

\item{hypo_color}{Color for hypo-methylated. (\bold{\emph{"#008800"}}).}

\item{mark_style}{Marker style for DMGs. (\bold{\emph{"point"}}, "line").}

\item{line_width}{Line width when \code{mark_style="line"}. (\bold{\emph{0.6}}).}

\item{line_height}{Line height relative to bar radius. (\bold{\emph{0.8}}).}
}
\value{
A \bold{\emph{ggplot object}} visualizing DMGs along chromosomes.
}
\description{
Plot \bold{\emph{DMGs hyper/hypo}} along chromosomes.
}
\examples{
# DMR results
dmr_file <- system.file(
  "extdata",
  "example.dmr",
  package = "GAnnoViz")

# Plot DMG expression
plot_dmg_exp(
  dmr_file = dmr_file,
  orientation = "horizontal",
  chrom_alpha = 0.1,
  chrom_color = "#008888",
  point_size = 1,
  point_alpha = 0.3,
  hyper_color = "#ff0000",
  hypo_color = "#008800",
  mark_style = "point",
  line_width = 0.6,
  line_height = 0.8)

}
\author{
benben-miao
}
