% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dmg_trend.R
\name{plot_dmg_trend}
\alias{plot_dmg_trend}
\title{Plot chromosomal DMGs trend}
\usage{
plot_dmg_trend(
  chrom_id,
  dmr_file,
  smooth_span = 0.1,
  hyper_color = "#ff000055",
  hypo_color = "#00880055",
  point_size = 3,
  point_alpha = 0.5
)
}
\arguments{
\item{chrom_id}{Chromosome identifier (e.g., \bold{\emph{"chr1"}}).}

\item{dmr_file}{DEG table from \bold{\emph{MethylKit}} analysis.}

\item{smooth_span}{Span for local regression smoothing. (\bold{\emph{0.1}}).}

\item{hyper_color}{Color for hyper-methylated DMRs. (\bold{\emph{"#ff0000"}}).}

\item{hypo_color}{Color for hypo-methylated DMRs. (\bold{\emph{"#008800"}}).}

\item{point_size}{Point size. (\bold{\emph{3}}).}

\item{point_alpha}{Point alpha. (\bold{\emph{0.5}}).}
}
\value{
A \bold{\emph{ggplot object}} chromosomal DMGs trend.
}
\description{
Plot chromosomal DMGs \bold{\emph{trend}}.
}
\examples{
# Example DMR table in GAnnoViz
dmr_file <- system.file(
  "extdata",
  "example.dmr",
  package = "GAnnoViz")

# Plot DMR trend
plot_dmg_trend(
  chrom_id = "chr1",
  dmr_file = dmr_file,
  smooth_span = 0.1,
  hyper_color = "#ff0000",
  hypo_color = "#008800",
  point_size = 3,
  point_alpha = 0.5
)

}
