% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_domains.R
\name{plot_gene_domains}
\alias{plot_gene_domains}
\title{Plot protein domains from Ensembl}
\usage{
plot_gene_domains(
  gene_name = NULL,
  species = "hsapiens",
  transcript_id = NULL,
  transcript_choice = "longest",
  palette = "Set 2",
  legend_ncol = 2,
  return_data = FALSE
)
}
\arguments{
\item{gene_name}{Official gene symbol. (\bold{\emph{"TP53"}}).}

\item{species}{Species in BioMart format. (\bold{\emph{"hsapiens"}}, "mmusculus").}

\item{transcript_id}{Ensembl transcript ID. (\bold{\emph{"ENST00000269305"}}).}

\item{transcript_choice}{Transcript choice. (\bold{\emph{"longest"}}, "canonical").}

\item{palette}{Color palette. (\bold{\emph{"Set 2"}}, "Set 2", "Set 3", "Warm", "Cold", "Dynamic", "Viridis", "Plasma", "Inferno", "Rocket", "Mako").}

\item{legend_ncol}{Legend columns per row. (\bold{\emph{2}}).}

\item{return_data}{Plot and data. (\bold{\emph{TRUE}}).}
}
\value{
Plot of \bold{\emph{protein domains}}.
}
\description{
Plot \bold{\emph{protein domains}} from Ensembl.
}
\examples{
# Plot TP53 domian
res <- plot_gene_domains(
  gene_name = "TP53",
  species = "hsapiens",
  transcript_id = NULL,
  transcript_choice = "longest",
  palette = "Set 2",
  legend_ncol = 2,
  return_data = TRUE)

res$plot
head(res$domain_data)

}
\author{
benben-miao
}
