% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snp_fst.R
\name{plot_snp_fst}
\alias{plot_snp_fst}
\title{Plot genomic weighted FST heatmap}
\usage{
plot_snp_fst(
  fst_file,
  bin_size = 1e+06,
  metric = "fst_mean",
  orientation = "horizontal",
  palette = c("#ffffff", "#aa00aa"),
  alpha = 0.9
)
}
\arguments{
\item{fst_file}{FST sliding window results. (\bold{\emph{CHROM, BIN_START, BIN_END, WEIGHTED_FST, N_VARIANTS}}.}

\item{bin_size}{Bin size in base pairs. (\bold{\emph{1e6}}).}

\item{metric}{Aggregation metric for bin fill. (\bold{\emph{"fst_mean"}}, "variant_count").}

\item{orientation}{Coordinate orientation. (\bold{\emph{"horizontal"}}, "vertical").}

\item{palette}{Continuous color for weighted FST. (\bold{\emph{c("#ffffff", "#aa00aa")}}).}

\item{alpha}{Tile alpha for heatmap. (\bold{\emph{0.9}}).}
}
\value{
A \bold{\emph{ggplot object}} showing genomic weighted FST heatmap.
}
\description{
Plot genomic \bold{\emph{weighted FST}} heatmap.
}
\examples{
# Example FST sliding window table in GAnnoViz
fst_file <- system.file(
  "extdata",
  "example.fst",
  package = "GAnnoViz")

# Plot weighted FST
plot_snp_fst(
  fst_file = fst_file,
  bin_size = 1e6,
  metric = "fst_mean",
  orientation = "horizontal",
  palette = c("#ffffff", "#aa00aa"),
  alpha = 0.9
)
}
\author{
benben-miao
}
