% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_ensembl_gff3.R
\name{retrieve_ensembl_gff3}
\alias{retrieve_ensembl_gff3}
\title{Download GFF3 Annotation from Ensembl}
\usage{
retrieve_ensembl_gff3(
  species = "mus_musculus",
  release = 115,
  only_chr = TRUE,
  dest_dir = tempdir()
)
}
\arguments{
\item{species}{Species name (e.g., "mus_musculus", "homo_sapiens").}

\item{release}{Ensembl release version (e.g., 115).}

\item{only_chr}{Logical. If \code{TRUE}, download only chromosomes (chr). Default \code{TRUE}.}

\item{dest_dir}{Destination directory. Default is session temporary directory.}
}
\value{
A \bold{\emph{character}} string of the downloaded file absolute path.
}
\description{
Download \bold{\emph{GFF3}} annotation file from Ensembl database.
}
\examples{
\dontrun{
# Download Human GFF3 (Chr only)
# gff_file <- retrieve_ensembl_gff3(
#   species = "homo_sapiens",
#   release = 115,
#   only_chr = TRUE,
#   dest_dir = tempdir())
}

}
\author{
benben-miao
}
