% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_ensembl_species.R
\name{retrieve_ensembl_species}
\alias{retrieve_ensembl_species}
\title{List Supported Species from Ensembl}
\usage{
retrieve_ensembl_species(release = 115)
}
\arguments{
\item{release}{Ensembl release version (e.g., 115).}
}
\value{
A \bold{\emph{data.frame}} with two columns: \code{Species} (formatted name) and \code{ID} (folder name).
}
\description{
List supported species and their IDs from Ensembl GFF3 FTP directory.
}
\examples{
\dontrun{
# List species for release 115
species_df <- retrieve_ensembl_species(release = 115)
head(species_df)
}

}
\author{
benben-miao
}
